Refactored code

+++ Easy/C/Ceil and Floor.c
@@ -5,5 +5,5 @@
 int main() {
     double X;
     scanf("%lf", &X);
-    printf("%" PRIu8 " %" PRIu8, (uint8_t) ceil(X), (uint8_t) floor(X));
+    printf("%" PRIu8 " %" PRIu8, ceil(X), floor(X));
 }
